merge: clean up overwrites from merge
