+++ web-registry/src/components/organisms/BuyCreditsModal/BuyCreditsModal.tsx
@@ -31,7 +31,7 @@ import {
 
 import { formatDenomText, microToDenom } from 'lib/denom.utils';
 
-import { UISellOrderInfo } from 'pages/Marketplace/Projects/Projects.types';
+import { UISellOrderInfo } from 'pages/Projects/Projects.types';
 
 import { BUY_CREDITS_MODAL_DEFAULT_VALUES } from './BuyCreditsModal.constants';
 import { useBuyCreditsModalStyles } from './BuyCreditsModal.styles';
