+++ web-registry/src/pages/Projects/Projects.config.ts
@@ -6,6 +6,6 @@ export const VIEW_ECOCREDITS = 'view your ecocredits';
 export const sortOptions = [
   { label: 'Price  - low to high', value: 'price-ascending' },
   { label: 'Price  - high to low', value: 'price-descending' },
-  { label: 'Credits available  - high to low', value: 'credits-descending' },
   { label: 'Credits available  - low to high', value: 'credits-ascending' },
+  { label: 'Credits available  - high to low', value: 'credits-descending' },
 ];
